+++ esp-hal/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `i2c::master::BusTimeout::Disabled` for ESP32-S2 (#3591)
 
 ### Changed